dagrr: Add end-to-end tests
Run them with:

    just dagrr

This ensures that 95% of the flags work as expected and the right
operations are happening behind the scenes. Meant to be run locally, and
results validated by a human. Running this in CI will be problematic
since contributions come via PRs, and they will **NOT** have access to
credentials such as the `FLY_API_TOKEN`. This works for now!

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>