Rename dagrr to dagr + other breaking changes
I made the `daggr` mistake too many times. If I couldn't instinctively
remember that it was `dagrr`, it means that the name wasn't right. Also,
double letters are a pain in general. `dagr` stands for `dag` (the heart
of Dagger) and `r` as in `runtime`.

Also, Connor was using it right. The `on-flyio deploy` interface was off.
I tweaked it a bunch & also moved the justfile in the right place so
that it is easier to get started with.

There is now a README which should capture how this is meant to be used.
Should make Connor happy.

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>